app/auth/jwt_handler.py
# physiobuddy_backend/app/auth/jwt_handler.py

from datetime import datetime, timedelta
from jose import JWTError, jwt

# Secret key to encode and decode JWT tokens
SECRET_KEY = (
    "your-secret-key"  # Replace this with something secure and hidden in production
)
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


# Create JWT token
def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


# Decode JWT token
def decode_access_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
