requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
ecdsa==0.19.1
    # via python-jose
fastapi==0.115.12
    # via -r requirements.in
idna==3.10
    # via anyio
passlib==1.7.4
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.1
    # via pydantic
python-jose==3.4.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via -r requirements.in
starlette==0.46.1
    # via fastapi
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
